[Add kernel_in_ubi feature to separate image format]
[Description]
Add kernel_in_ubi feature to separate image format
1. kernel_in_ubi will decide partition layout and enable overlayfs
on/off

[Release-log]

+++ recipes-devtools/initrd/systemd_%.bbappend
@@ -5,11 +5,13 @@ FILES_${PN} += " /overlay \
                /rom \"
 
 do_install_append() {
+       if ${@bb.utils.contains('DISTRO_FEATURES','kernel_in_ubi','true','false',d)}; then
        install -d ${D}/overlay
        install -d ${D}/rom
        install -m 0755 ${WORKDIR}/init_readonlyfs.sh ${D}${rootlibexecdir}/init_readonlyfs.sh
        [ -e ${D}/sbin/init ] && rm -rf ${D}/sbin/init
        ln -s ${rootlibexecdir}/init_readonlyfs.sh ${D}/sbin/init
+       fi
 }
 
 FILES_${PN} += " ${rootlibexecdir}/init_readonlyfs.sh "
